Clean and delete API project for now
