--- src/MangaNew.Core/Services/IMangaProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace MangaNew.Core.Services
-{
-    public interface IMangaProvider
-    {
-        Task<IReadOnlyList<MangaRelease>> GetLatestMangaAsync(CancellationToken cancellationToken = default);
-    }
-}
