+++ src/MangaNews.Core/MangaNews.Core.csproj
