+++ src/MangaNews.Core/Services/IMangaProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace MangaNews.Core.Services;
+
+public interface IMangaProvider
+{
+    Task<IReadOnlyList<MangaRelease>> GetLatestMangaAsync(CancellationToken cancellationToken = default);
+}
