+++ src/MangaNews.Parser/MangaNews.Parser.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MangaNew.Core\MangaNew.Core.csproj" />
+    <ProjectReference Include="..\MangaNews.Core\MangaNews.Core.csproj" />
   </ItemGroup>
 
 </Project>
